feat: Separate caip10-link handler from caip10-link doctype
