+++ packages/core/package.json
@@ -32,6 +32,7 @@
     "@ceramicnetwork/common": "^0.15.4",
     "@ceramicnetwork/docid": "^0.4.2",
     "@ceramicnetwork/doctype-caip10-link": "^0.12.6",
+    "@ceramicnetwork/doctype-caip10-link-handler": "^0.12.6",
     "@ceramicnetwork/doctype-tile": "^0.12.8",
     "@ceramicnetwork/doctype-tile-handler": "^0.12.8",
     "@ceramicnetwork/pinning-aggregation": "^0.1.9",
