+++ packages/doctype-caip10-link-handler/README.md
@@ -1,0 +1,15 @@
+# Ceramic Account Link doctype handler
+
+> This package contains the Caip-10 Link doctype handler implementation.
+
+### Installation
+```shell
+npm install @ceramicnetwork/doctype-caip10-link
+```
+
+## Contributing
+We are happy to accept small and large contributions. Make sure to check out the [Ceramic specifications](https://github.com/ceramicnetwork/specs) for details of how the protocol works.
+
+## License
+
+Apache-2.0 or MIT
