+++ packages/doctype-caip10-link-handler/babel.config.js
@@ -1,0 +1,10 @@
+// babel.config.js
+module.exports = {
+  presets: [
+    ['@babel/preset-env', {targets: {node: 'current'}}],
+    '@babel/preset-typescript',
+  ],
+  plugins: [
+    ["@babel/plugin-proposal-decorators", { "legacy": true }],
+  ]
+};
