+++ packages/doctype-caip10-link-handler/src/index.ts
@@ -1,0 +1,1 @@
+export * from './caip10-link-doctype-handler'
