+++ packages/doctype-caip10-link/src/index.ts
@@ -1,2 +1,1 @@
 export * from './caip10-link-doctype'
-export * from './caip10-link-doctype-handler'
